Implement IDE APIs diff builder used to build "available since" external annotations necessary for IDEA-201007
The diff builder is used to track introduction of new APIs in the IntelliJ Platform. The versions of IDEs when APIs were first introduced are recorded in form of external annotations, where 'org.jetbrains.annotations.ApiStatus.AvailableSince' specifies the first versions. The annotations roots are packaged to .jar archives and uploaded to the IntelliJ artifacts repository, from where they are downloaded by DevKit side and included to the project. A corresponding inspection tracks versions when APIs, which are used in plugins, were first introduced and compares that with [since; until] compatibility range specified in plugin. If the range has some versions not covered by the used APIs, a warning is emitted.